Install the .net core SDK into a temporary c:\temp\tools directory to mimic tech suite
AA previously did not automatically install the SDK and relied soley on tech
suite to install this prereq. This version of the AA installer will ship by
itself, without the tech suite installer which will ship afterwards. In order
to provide a similar behavior, and because we can call install-dotnetcore
many times with the same arugment path, the decision was made to install the
prereq in both AA and tech suite installer