Include basic install.ps1 entry point script
install.ps1 will not be considered by nuget as an automatic script
and won't run when this package is installed. See nuget pack output
warning below for more details

WARNING: NU5110: The script file 'install.ps1' is outside the 'tools'
folder and hence will not be executed during installation of this
package. Move it into the 'tools' folder.